+++ homeassistant/components/plex/const.py
@@ -46,6 +46,7 @@ X_PLEX_VERSION = __version__
 AUTOMATIC_SETUP_STRING = "Obtain a new token from plex.tv"
 MANUAL_SETUP_STRING = "Configure Plex server manually"
 
+SERVICE_PLAY_ON_OTHER = "play_media"
 SERVICE_PLAY_ON_SONOS = "play_on_sonos"
 SERVICE_REFRESH_LIBRARY = "refresh_library"
 SERVICE_SCAN_CLIENTS = "scan_for_clients"
