+++ homeassistant/components/plex/services.yaml
@@ -1,5 +1,19 @@
+play_media:
+  description: Play music hosted on a Plex server on a `media_player` from a supported integration.
+  fields:
+    entity_id:
+      description: Entity ID of a supported `media_player`.
+      example: "media_player.sonos_living_room"
+    media_content_id:
+      description: The ID of the content to play. See https://www.home-assistant.io/integrations/plex/#service-play_media for details.
+      example: >-
+          '{ "library_name": "Music", "artist_name": "Stevie Wonder" }'
+    media_content_type:
+      description: The type of content to play. Must be one of movie, music, episode, or playlist.
+      example: "music"
+
 play_on_sonos:
-  description: Play music hosted on a Plex server on a linked Sonos speaker.
+  description: DEPRECATED. Use `plex.play_media`. Play music hosted on a Plex server on a linked Sonos speaker.
   fields:
     entity_id:
       description: Entity ID of a media_player from the Sonos integration.
